+++ lib/src/core/utils/js_compat_web.dart
@@ -1,3 +1,0 @@
-/// Web implementation that simply returns the provided function.
-F allowInterop<F extends Function>(F f) => f;
-